+++ CHANGELOG.md
@@ -3,6 +3,8 @@
   attempting to restore the machine state.
 * Added Clang support.
 * Added Arm Linux GCC support.
+* Added Conan 2.0 support.
+* Replace Utils/base64.hpp with Conan package libbase64.
 
 1.5.1 [22/04/16]
 * Fixed api compatibility break in 1.5.0.